Inlined PullDerived/PullOps/pull1 in to Pull class/object
